Add READMEs explaining the fuzzball-synth directory
Note that I'm adding these while it's currently under discussion,
even though we've decided we do not plan to include this material in
upcoming deliverable because the UMN licensing issues are not worked
out. I hope this context should be somewhat useful even to internal
users, and we can have it ready when we do deliver the code.