Optimizations and a few extra examples in FuzzBALL synth.
Specify -branch-preference to prefer either working adaptors or
counterexamples, as required.
Rewrite the wrapper code in assembly to enforce use of only conditional
moves instead of branches.
Add two more examples, a small variant of low-bit for which equivalence
fails, and a harder example that's more like a classic synthesis problem.
Add a better framework for choosing the FuzzBALL and STP binary
locations.
Make the output a bit less verbose.