Initialize some more malloc and stdio state
Special casing various behaviors of the dynamic linker, which we may
want to do more elegantly in the future. Based on some testing with
getpass(3), I think this should address a number of false negatives,
though perhaps in exchange for false positives.
Copy the quoting of command outputs for the benefit of cut and paste.