Disable C.E. cache for struct counterexample generation
When the path condition is very long, filtering the set of satisfying
assignments every time we add a branch condition is too slow. For the
future we might want to have a version of the working cache that is
only checked right at a branch, to get the core benefit of that
optimization at less cost.

Add a script that counts up how much of a FuzzBALL SMT file is
used by each variable. I used this to debug the problem with the
files getting bigger under -split-target-formulas.