Handling noop-ed syscalls in conc. adaptor search
CE search will create a file named "noop_syscalls.lst" and populate it
with the syscall number for every syscall that it noops.
Concrete adaptor search will then read this file, if it exists, and
replace the return value (RAX) with 0 for every noop-ed system call
during f1 and f2.

This does not noop the system call, but tries to match the return
value of noop-ed system calls as done in FuzzBALL.