Synthesizing the adapters between the RC4 encryption functions using a one byte-long symbolic input string.
We make FuzzBALL use a one-byte symbolic input string by not
zero-initializing or setting to another concrete value the first
byte of the input string. Instead we pass the counter-example in the
first byte of the input string as a fuzzball_extra_arg from C.E. search.
We also increase the input_string_addr by 2*region-size at the end of
every C.E. search step to create a new input string address for the next
C.E. search